pyproject.toml
[project]
name = "audio-transcript"
version = "0.1.0"
description = "YouTube video to speaker-diarized transcript"
requires-python = "==3.11.9"
authors = [
    {name = "Luis Alejandro Bordo García", email = "bgluiszz@gmail.com"},
]
dependencies = [
    "openai-whisper",
    "pandas",
    "pyannote.audio",
    "pyannote.core",
    "pyannote.database",
    "pyannote.metrics",
    "pyannote.pipeline",
    "pydantic",
    "python-dotenv",
    "pydub",
    "torchaudio",
    "torchvision",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "black",
    "pytest",
    "jupyter",
    "nbqa",
    "pandas-stubs",
]

[project.scripts]
audio-transcript = "audio_transcript.pipeline:cli"
present-transcript = "audio_transcript.present_transcript:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "pyannote.*",
    "whisper",
    "torchaudio",
    "pydub.*",
    "yt_dlp",
]
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ['py311']
skip-string-normalization = true
